[z0273] Send looted creature guid in proper field of loot packets.
Also use ObjectGuid in related functions.


(based on commit cee592b)

(based on commit aeb70d1)

+++ src/game/ObjectMgr.cpp
@@ -379,7 +379,7 @@ void ObjectMgr::LoadPointOfInterestLocales()
             int idx = GetOrNewIndexForLocale(LocaleConstant(i));
             if(idx >= 0)
             {
-                if(data.IconName.size() <= idx)
+                if(data.IconName.size() <= size_t(idx))
                     data.IconName.resize(idx+1);
 
                 data.IconName[idx] = str;
